requirements.txt
# --- EL CEREBRO (LangGraph y LLM) ---
langgraph>=0.0.65
langchain-core>=0.2.0
langchain-anthropic>=0.1.15

# --- EL CUERPO (Navegador Remoto) ---
playwright>=1.44.0
Pillow>=10.2.0

# --- LA BOCA Y LOS OÍDOS (Videollamada en Tiempo Real) ---
livekit-agents>=0.6.0
livekit-api>=0.5.0
livekit-plugins-openai>=0.5.0

fastapi
uvicorn
pydantic